IMPORTANT ISSUE RESOLVED
i woke up at 3am bc i had a nightmare something might be broken. 

i noticed visiting the deployed link in incognito mode as a first time user loaded a blank page instead of rendering the index.html template. users would have to refresh. that is annoying. SO THIS IS HOW IT IS FIXED:

when you generate a new user_id and return the response (resp), it stops the request from continuing to the intended endpoint (index). that’s why the page is blank until a refresh happens...on the first request, the cookie is set, but the page content isn't rendered. 

all that is done differently here is explicitly setting the cookies within the endpoint.

committing to main bc yolo

Signed-off-by: Cameron Copenhaver <151974261+ccopenh1@users.noreply.github.com>