Update detected.html
Fix issue where clicking the `X` only closes the modal and leaves the user still on the `detected` page but misleadingly appears to be on the `home` page, which would result in an error if the user attempted to record and upload another song.

Clicking the `X` now has an `onclick` event handler to ensure the user is properly redirected to the home page and the modal disappears. Now the user is able to record another song for detection!

Signed-off-by: Cameron Copenhaver <151974261+ccopenh1@users.noreply.github.com>